review(pool,server): split oversized files by responsibility
pool.go 698->452: Reconcile+reclaimOrphanArchiveCks -> reconcile.go,
SetPools+normalizePoolSpec -> setpools.go (matching the package's
operation-per-file convention). server.go 580->490: request-context,
body-decode, and response helpers -> server_http.go. Pure moves, no
behavior change.
