+++ sandboxd/pool/pool.go
@@ -318,154 +318,6 @@ func (m *Manager) Run(ctx context.Context) {
 	}
 }
 
-// Reconcile aligns state after a daemon restart: re-adopt persisted claims
-// whose VMs are still running (or hibernated), drop the rest, and remove any
-// sbx-prefixed VM nobody owns (stale pool VMs and golden builders from a
-// previous life). It must run once at startup, before the server: it swaps
-// in fresh records, which would bypass in-flight Transition locks.
-func (m *Manager) Reconcile(ctx context.Context) error {
-	claims, err := m.store.load()
-	if err != nil {
-		return err
-	}
-	vms, err := m.eng.List(ctx)
-	if err != nil {
-		return fmt.Errorf("list vms: %w", err)
-	}
-	live := make(map[string]types.VMRecord, len(vms))
-	for _, vm := range vms {
-		live[vm.Config.Name] = vm
-	}
-	snaps, snapsErr := m.eng.SnapshotList(ctx)
-
-	owned := map[string]bool{}
-	referenced := map[string]bool{}
-	m.mu.Lock()
-	for id, sb := range claims {
-		rec, ok := live[sb.VMName]
-		switch {
-		case sb.ArchiveCk != "":
-			// Archived: no local VM by design; the store ck is the durable
-			// state. Adopt the stub so the id/token survive restart and the
-			// first exec wakes it (wakeArchived).
-			m.claimed[id] = sb
-			m.tenantDelta(sb.Tenant, 1)
-			continue
-		case ok && rec.State == vmStateRunning:
-			sb.VsockSocket = rec.VsockSocket
-			// Running with either flag set = a wake crashed between restore
-			// and commit, or a hibernate intent never reached the engine;
-			// clearing them un-bricks the claim and unreferences the
-			// snapshot for the sweep below.
-			sb.HibernateSnap, sb.PendingSnap = "", ""
-		case ok && sb.HibernateSnap != "":
-			// Hibernated: the VM is stopped by design and wakes on demand.
-			sb.PendingSnap = ""
-		case ok && sb.PendingSnap != "" && (snapsErr != nil || slices.Contains(snaps, sb.PendingSnap)):
-			// Stopped under a journaled hibernate intent whose commit never
-			// landed: the intent names the wake image, adopt it. A verified-
-			// missing image means the hibernate never completed — fall
-			// through and drop; an unverifiable list adopts (a failed wake
-			// beats a destroyed claim).
-			sb.HibernateSnap, sb.PendingSnap = sb.PendingSnap, ""
-		default:
-			continue
-		}
-		m.claimed[id] = sb
-		m.tenantDelta(sb.Tenant, 1)
-		owned[sb.VMName] = true
-		referenced[sb.HibernateSnap] = true
-	}
-	saveErr := m.store.save(m.claimed)
-	for _, p := range m.pools {
-		dir := filepath.Join(m.goldensDir(), p.key.Hash())
-		if dirExists(dir) {
-			p.goldenDir = dir
-		}
-	}
-	m.mu.Unlock()
-	// A crash mid-export leaves a *.tmp staging dir no build or promote of
-	// this life will reuse.
-	tmps, _ := filepath.Glob(filepath.Join(m.goldensDir(), "*.tmp"))
-	if err := m.ckpts.SweepStaging(); err != nil {
-		log.WithFunc("pool.Reconcile").Error(ctx, err, "sweep checkpoint staging")
-	}
-	if err := m.tpls.SweepStaging(); err != nil {
-		log.WithFunc("pool.Reconcile").Error(ctx, err, "sweep template staging")
-	}
-	m.migrateLegacyTemplates(ctx)
-	for _, tmp := range tmps {
-		_ = os.RemoveAll(tmp)
-	}
-
-	logger := log.WithFunc("pool.Reconcile")
-	m.reclaimOrphanArchiveCks(ctx, claims)
-	var stale []string
-	for name := range live {
-		if strings.HasPrefix(name, vmPrefix) && !owned[name] {
-			stale = append(stale, name)
-		}
-	}
-	m.runBounded(ctx, len(stale), func(ctx context.Context, i int) {
-		m.destroy(ctx, stale[i])
-		logger.Infof(ctx, "removed stale VM %s", stale[i])
-	}).Wait()
-
-	// Snapshot sweep, symmetric to the VM sweep: a hibernate snapshot no
-	// adopted claim references is an orphan, and fork/golden-build snapshots
-	// are transient by construction — none can span a restart. A list failure
-	// only skips the sweep: GC must not brick startup.
-	if snapsErr != nil {
-		logger.Warnf(ctx, "snapshot sweep skipped: %v", snapsErr)
-	} else {
-		var orphans []string
-		for _, snap := range snaps {
-			orphanHib := strings.HasPrefix(snap, hibernatePrefix) && !referenced[snap]
-			if orphanHib || strings.HasPrefix(snap, forkPrefix) || strings.HasPrefix(snap, goldenPrefix) {
-				orphans = append(orphans, snap)
-			}
-		}
-		m.runBounded(ctx, len(orphans), func(ctx context.Context, i int) {
-			m.dropSnap(ctx, orphans[i])
-			logger.Infof(ctx, "removed orphan snapshot %s", orphans[i])
-		}).Wait()
-	}
-	now := time.Now()
-	for _, sb := range m.claimed {
-		sb.TouchAt(now)
-	}
-	logger.Infof(ctx, "adopted %d claims, %d VMs live", len(m.claimed), len(live))
-	return saveErr
-}
-
-// reclaimOrphanArchiveCks reaps archives that crashed between their
-// checkpoint publish and the journal commit: the flag hides them from
-// listings and deletes, so only the journal identifies ours — the sandbox is
-// in it under a different (or no) ArchiveCk.
-func (m *Manager) reclaimOrphanArchiveCks(ctx context.Context, claims map[string]*types.Sandbox) {
-	logger := log.WithFunc("pool.reclaimOrphanArchiveCks")
-	metas, err := m.ckpts.Metas(ctx)
-	if err != nil {
-		logger.Warnf(ctx, "list archive cks skipped: %v", err)
-		return
-	}
-	for _, raw := range metas {
-		var ckpt types.Checkpoint
-		if json.Unmarshal(raw, &ckpt) != nil || !ckpt.Archive {
-			continue
-		}
-		orig, mine := claims[ckpt.SandboxID]
-		if !mine || orig.ArchiveCk == ckpt.ID {
-			continue
-		}
-		if err := m.deleteCkLocked(ctx, ckpt.ID); err != nil {
-			logger.Warnf(ctx, "reclaim %s: %v", ckpt.ID, err)
-		} else {
-			logger.Infof(ctx, "reclaimed orphan archive ck %s", ckpt.ID)
-		}
-	}
-}
-
 // FlushClaims synchronously persists the current claim set — the shutdown
 // hook that closes the window where a detached recommit has not converged yet.
 func (m *Manager) FlushClaims() error {
@@ -482,91 +334,6 @@ func (m *Manager) untrack(set map[string]struct{}, key string) {
 	m.mu.Lock()
 	delete(set, key)
 	m.mu.Unlock()
-}
-
-// SetPools replaces the node's desired warm targets. Existing claims are not
-// affected; only unclaimed warm VMs are trimmed or refilled.
-func (m *Manager) SetPools(ctx context.Context, specs []config.PoolSpec) error {
-	desired := make(map[types.PoolKey]config.PoolSpec, len(specs))
-	hashes := make(map[string]types.PoolKey, len(specs))
-	for _, spec := range specs {
-		spec = normalizePoolSpec(spec)
-		if err := m.validate(spec.PoolKey); err != nil {
-			return err
-		}
-		if err := spec.ValidateLimits(); err != nil {
-			return fmt.Errorf("%w: %v", ErrBadCount, err)
-		}
-		// The manager neither validates nor keeps egress policy yet;
-		// accepting it here would silently drop it.
-		if spec.Egress != nil {
-			return fmt.Errorf("%w: pool %q: egress is set in the config file, not via the API", ErrBadKey, spec.Template)
-		}
-		if existing, ok := hashes[spec.Hash()]; ok && existing != spec.PoolKey {
-			return fmt.Errorf("%w: pool key hash collision between %q and %q", ErrBadKey, existing.Template, spec.Template)
-		}
-		if _, ok := desired[spec.PoolKey]; ok {
-			return fmt.Errorf("%w: duplicate pool %q", ErrBadKey, spec.Template)
-		}
-		hashes[spec.Hash()] = spec.PoolKey
-		desired[spec.PoolKey] = spec
-	}
-
-	var trim []string
-	m.mu.Lock()
-	now := time.Now()
-	for key, p := range m.pools {
-		spec, ok := desired[key]
-		if !ok {
-			p.floor = 0
-			p.warmMax = 0
-			p.idle = 0
-		} else {
-			p.applySpec(spec)
-		}
-		target := p.effectiveTarget(now)
-		for len(p.warm) > target {
-			n := len(p.warm) - 1
-			trim = append(trim, p.warm[n].VMName)
-			p.warm = p.warm[:n]
-		}
-		if !ok && !p.building && p.refilling == 0 {
-			delete(m.pools, key)
-		}
-	}
-	for key, spec := range desired {
-		if p := m.pools[key]; p != nil {
-			continue
-		}
-		p := &pool{key: key}
-		p.applySpec(spec)
-		// A golden already on disk (from this pool's earlier life) is
-		// adopted; buildGolden covers the rest.
-		if g := filepath.Join(m.goldensDir(), key.Hash()); dirExists(g) {
-			p.goldenDir = g
-		}
-		m.pools[key] = p
-	}
-	// Recompute rather than latch: removing every idle pool turns the
-	// sweep off again.
-	m.idleEnabled = m.idleDefault > 0
-	m.archiveEnabled = m.archiveAfterDefault > 0
-	for _, p := range m.pools {
-		if p.idle > 0 {
-			m.idleEnabled = true
-		}
-		if p.archiveAfter > 0 {
-			m.archiveEnabled = true
-		}
-	}
-	m.mu.Unlock()
-
-	runCtx := context.WithoutCancel(ctx)
-	for _, name := range trim {
-		m.destroy(runCtx, name)
-	}
-	m.refillOnce(runCtx)
-	return nil
 }
 
 // SetTemplateNotifier wires the immediate-republish hook; call it before
@@ -640,19 +407,6 @@ func (m *Manager) validate(key types.PoolKey) error {
 
 func (m *Manager) goldensDir() string {
 	return filepath.Join(m.dataDir, "goldens")
-}
-
-// normalizePoolSpec fills the wire defaults, mirroring ClaimRequest.Key():
-// API requests default net/size; config files stay explicit (Validate
-// rejects empty there).
-func normalizePoolSpec(spec config.PoolSpec) config.PoolSpec {
-	if spec.Net == "" {
-		spec.Net = types.NetNone
-	}
-	if spec.Size == "" {
-		spec.Size = types.SizeSmall
-	}
-	return spec
 }
 
 // newStoreView builds one id-namespaced view of the configured backend.
